Add `msw` handler for query to DDN
This will get more complicated with more queries. One interesting
challenge is that we can't rely on the URL for request matching,
since it's always the same URL, but with a varied SQL body. This
is similar to GraphQL and therefore the msw graphql request matcher,
except that we're using SQL instead. So we would need to write a
custom request matcher to route based on query. Until then, we can
just conditionally return "pass through" to move to the next request
handler.