Add `yarn codegen` to generate typings from splitgraph importer plugins
Generate types for `CredentialsSchema`, `ParamsSchema`, and `TableParamsSchema`
for each plugin in the `externalPlugins` dump (just a JSON file for testing atm)

Using: https://github.com/bcherny/json-schema-to-typescript

It works pretty well, but has some trouble with `items[]` (and/or our schema
is invalid). It still runs, but emits `unknown` in these cases, and we can
ignore the typechecking error

Don't check the generated code into git, for now. Arguably this is equivalent
to a schema and should be checked into the repo, but it's codegen, which in the
past we've done as a (pre-)build step