Add method `fetchAccessToken` to exchange API key/secret for token
The `X-API-Key` and `X-API-Secret` headers are only supported by
Splitgraph for the DDN query endpoint (where we send SQL). For
the GraphQL API, an access token is required.

Therefore, add a methd to `db-splitgraph` to exchange a keypair
for an access token (is this the best place for it? maybe)