Fix types during build stage, set jsx = `react-native`
- Add missing references
- Delete extraneous references
- Maybe temporary: During build stage of `@madatdata/react`,
set `jsx = react-native` in `tsconfig.build.json`, which is
the same as `preserve` but does not change the file extension
to `.jsx`. This is necessary because `tsc-multi` doesn't know
to rename `.jsx` to `.mjsx` or similar, so it leaves it as `.jsx`
but rewrites the import from non-existent `.mjs`. That could be
fixed, but it's possible that `react` might be the right way
to do it anyway