Add `@madatdata/react` as dependency of example (maybe works?)
Want to specify using `latest` identifier, so user can copy example,
instead of needing to specify `workspace:*`. So specify `latest` in
the example but then resolve it to `workspace:*`.

Since `examples` has `nmHoistingLimits: workspace`, this means that there
are two instances of `@madatdata/react`, which might imply that it's
necessary to `yarn install` for the example to get the latest upgrade
(not to mention it will expect that package's `build` to be latest)