Change import from node:webcrypto to webcrypto
I think in Node 16 it needs to be node:webcrypto, but
this at least works in Node 18 and fixes a bug in the
built React package where Webpack (as used by the Next.js
example) is unable to compile node:webcrypto.

But this has also changed back and forth so it might
not be right yet.