Import types in generated files from merged type file
There are some types we want to manually craft
or augment in the GraphQL API, and ideally we
shouldn't need to import them separately from the
generated types.

Instead of directly importing from the generated
schema file, import from a barrel file (that is
checked into the repository) which exports from
_both_ the generated types file, and the new
`explicit-types.ts`. This way, the `.generated.ts`
files will all start with an import from `unified-types.ts`