Inject seafowlClient into plugins from db calling the plugin
Seafowl needs a client in order to execute DDL commands,
and operations like importing data are done through the
same HTTP API as making queries. This is different from
db-splitgraph which uses the GraphQL API (and thus graphql-api-client)
as the management interface for importing and exporting data.

Usually the client and database are defined in parallel
while setting up a data context, but in some limited cases,
the database needs access to a client (like Seafowl import).

However, the client depends on the db-seafowl class being
instantiated to an object (or at least, it depends on enough
properties of it that I don't want to untangle exactly what
can be decoupled into a static method). So, we need to allow
plugins to be instantiated with an empty seafowlClient, but
later cloned via `withOptions({ seafowlClient: ... })`, creating
an effect where every plugin has a parent instance that doesn't
interact with any client, and can spawn child instances that do.