Export test helpers for makeDb of both splitgraph and seafowl
In the core package, we test at the highest level of abstraction
(closest to the user), and we have tests for both seafowl and
splitgraph `makeDb` functions. But now we want to test them together
(export data from splitgraph, import it into seafowl) so we need to
add a new test for that, but it needs to do the same repetitive
initializing of each database as the individual test files. So for
now, just export those helpers with the convention that `makeRealDb`
refers to an integration test and `makeDb` refers to a mocked one
for unit tests (even though this isn't true for seafowl atm)