Make non-integration seafowl test work without credential
Some Seafowl tests already run in CI, if they don't depend
on the actual databaes being up. But in CI there is no
credential, which means that the seafowl data context will
serialize differently than when there is one. (Note: using
an anonymous credential like this may not be supported yet,
since it just hardcodes anonymous-token and the fetch implementation
will transform the headers to add that to the Authorization header,
but this at least makes the tests pass in CI)