Add EventHandler implementing "Target" (emitter) and "Listener" (handler)
This is ported from Deno code in an earlier experimental repository.  The
source code required no changes. The test code required changing to use `await`
and `Promise` instead of `Deferred` from the Deno tests

It implements an "EventHandler," which might not be the best name, because the
same object can be either/or/both a "handler" (listens for events) and a
"target" (emits events, is source of event processed by handler)

The test shows awaiting an event to be dispatched, _and_ for its handlers to
resolve. This is done by passing a `resolve` argument to the dispatch function,
which is the reason for awaiting an async function wrapped with an outer
Promise. In Deno this is a bit more natural because the `resolve` argument can
be set to the return value of `deferred()`. Absent a `deferred` polyfill, it's
easier to wrap the call to the dispatch function with a promise. And since it's
optional, one might argue it's also cleaner - the inner promise is awaiting the
dispatch to complete; whereas the outer (optional) promise is awaiting all
subscribed handlers to that dispatch to complete.