Upgrade esbuild and rewrite its build scripts after major breaking changes
We use esbuild in tiny-react-sandbox, which we use so that we can
test the react code without needing to go through the whole process
of building packages, publishing to Verdaccio and building examples
on each change.

We need to upgrade esbuild (or at least we thought we needed to
when this started two hours ago) to fix issues importing from crypto
package (which we use for Seafowl hashing).

The latest ESBuild contains breaking changes because it rewrites
the incremental build APIs, so we need to rewrite basically the
entirety of dev.mjs to use the new APIs, which also include live
reload so we can drop our hacky implementation of it.

This commit implements dev.mjs to work properly with watch mode,
live reload, and also importing from 'crypto'

See: https://github.com/evanw/esbuild/pull/2816