Move from `jsx: preserve` to `jsx: react-jsx` in `tsconfig.base.json`
Without this change, files in tiny-react-playground only transform
TSX for direct entries in that project, so importing from `@madatdata/react`
causes syntax errors. With this change, the JSX is transformed properly.

Typechecks pass, but there's a slight possibility this breaks the
built packages (i.e. those consumed by examples, and real packages).
If this turns out to be the case then we will need to thread through
another tsconfig variant for esbuild.