Append .csv extension to all Seafowl GET queries
This is necessary for Cloudflare to cache the response and also for
Cloudflare to add cache-control headers for the browser to cache
the response. Without the extension, neither Cloudflare nor the
browser will cache the response.

The actual extension used is meaningless, but must be one recognized
by Cloudflare as a cacheable asset. The returned data is still in
jsonlines format, even with a .csv extension.