Re-export `@madatdata/core` from `@madatdata/react`
This might be a bad idea, but it means React consumers only
need to install one package. The point of this is so that we
can get rid of the duplicative /hooks and /hooks-seafowl, since
that's a bad pattern and it's also causing issues with JSX building
in the downstream packages that expect only one entrypoint. We
want to instead let the consumer inject which database they want
to use into the hook context.