Drop `--coverage` from `test-ci` to hopefully eliminate weird hanging
There continues to be some issue in GitHub actions where vitest
will hang after completing, because apparently some file handle
is not being closed. This commit drops the `--coverage` flag under
the assumption that it's somehow causing the bug. We can bring it
back later after upgrading vitest and the GHA runner to hopefully
fix whatever is the underlying issue.