Upgrade to TypeScript 5.0, get typechecking working but build still broken
Need to make some decisions about whether/how to emit CJS modules,
because as long as we have verbatimModuleSyntax enabled, we cannot
use the typical `export const = "x"` format, and tsc-multi rewriting
is also now breaking for this on v5, and patching it to do the rewrites
may be non trivial since it can involve shifting large blocks of code.

Alternatively we could just not opt into the flag, in which case
we should also restore the `isolatedModules` flag (which was removed
as redunant).

See:

* https://devblogs.microsoft.com/typescript/announcing-typescript-5-0/#verbatimmodulesyntax
* https://github.com/microsoft/TypeScript/pull/52203
* https://github.com/tommy351/tsc-multi/issues/19