Include `dbname` in constructed seafowl query URL, improve guessing algo, and add test
In the `makeQueryURL` strategy of `db-seafowl.ts`, always include
the dbname prefix, even if it's `default`, since that is the same
as not including the prefix.

Improve the hacky guessing algorithm for choosing between POST
or GET based on the content of the query: if it starts with a DDL
or DML statement, then assume POST. Otherwise, assume GET, but
if there is no SELECT found, print a warning to the user.

Add baisc unit tests for `makeQueryURL`