Add `zx` to repository, and add command e.g. `yarn zx scripts/print-cwd.ts`
Support running the command from the root of the repository or
from within the `scripts` workspace. This doesn't use `yarn workspace run`
because that would change the CWD, which breaks esno and zx magic. So
whatever directory you're in when you call `yarn zx` will be the CWD
as seen by the executing script. For this reason, for consistency
it's preferred to run scripts from the root.