Fix field parsing from content-type header and test it against live build
See: https://github.com/splitgraph/seafowl/pull/367

Parse fields from content-type header (we still aren't munging the arrow
schema to our own shape), and get it working with a live build of the
new feature.

Important note: The value of the HTTP parameter is an _un_-escaped JSON
string inside double quotes, by our convention, so we can just slice the
double quote off the beginning and end of the parameter value and then
parse it as JSON.

This commit adds some live tests, but they need a server with the new
build to work, and they're integration tests anyway (which wouldn't run
in CI), but they serve primarily as a place to collect the shape of
responses for later accurate mocking.

It also updates the existing mocked tests to use the new scheme (in particular,
the value is not escaped inside the double quotes).