Make strategy functions somewhat overrideable
Every client implementation (provided by a db implementation) must
define all of them or none of them, but sometimes (like when defining them
for tests), we don't want the end consumer to override the functions
provided by the db implementation. Work around that by making the functions
provided by the db implementation call the one provided in the options
if it exists. For some strategies, allow returning null to defer to the
default implementation.

Add comments to remind self this is a hacky mess.

Also: Add URL encoding of X-Seafowl-Query header before sending,
but after fingerprinting