Remove experimental Arrow code and packages
This branch started with some experiments using the `apache-arrow`
package, and we'll eventually bring those back, but the package
is not necessary for current functionality (we convert Arrow fields/types
to SQL fields/types with basic JS, and don't do anything fancy involving
the Arrow library itself).

Accordingly, delete that code for now.