Add `abortSignal` option to `client.execute` method
Support an `abortSignal` in the `client.execute` method, which will
abort the query when passed. Currently it's only implemented for the
HTTP client, i.e. not the Postgres client; and it's up to each client
whether/how to implement it. For the HTTP client, when a request is
aborted, the error is caught (like any other network error would be),
and `result.error` includes a `name` property which is `AbortSignal`,
taken from `DOMException["name"]` (consistent with Web standard behavior).