Support alternative function signature for `makeSeafowlDataContext`
This just takes a string, and optionally, a second argument that is
a simplified options object containing only `dbname`. Example:

```ts
const { client } = makeSeafowlHTTPContext("https://demo.seafowl.cloud", { dbname: "bob" })
```

This doesn't use function overloads because there were some issues
with "overloads not externally visible" that I didn't want to deal with.

This is a hack. But it offers a simplified API without mucking about
with all the generic options arguments for now.