Return instance of `Object.getPrototypeOf(this)` instead of storing `this.DerivedClass`
In the `withOptions` builder method, which is defined in the abstract
base class but which returns an instance of the derived class (with
an obviously unknown name), return an instance of the derived class
by referencing its constructor with `Object.getPrototypeOf(this).constructor`,
instead of forcing the derived class to store a reference to itself in
a property variable. This is arguably a less hacky method, but still
depends on the assumption that the derived class does not override
the constructor (but if it does, it can always override the builder
method as well).