Add `SplitgraphExportToSeafowlPlugin` for exporting from Splitgraph to Seafowl
We already supported "manually" exporting from Splitgraph to Seafowl
by first exporting a `.parquet` file via the existing `ExportQueryPlugin`,
and then importing it into Seafowl with `SeafowlImportFilePlugin`.

However, the Splitgraph API supports the more robust use case of
exporting one or multiple queries, tables and VDBs from Splitgraph
to Seafowl, via the GraphQL mutation `exportToSeafowl` which is
available with an authenticated request to the GraphQL API.

This commit adds a plugin to call that API with one or multiple queries,
tables or VDBs (with the types being fairly raw and maybe a bit cumbersome,
albeit still equipped with full autocomplete and sufficient TypeDoc annotations),
and to wait for the tasks to resolve before returning an object containing
the result, including a list of passed jobs, and if applicable, a list of
any failed jobs (since each export target creates a separate job, we need
to wait for them separately, and each can fail or pass independently of the others).

The result type is a bit of a mess, but it doesn't actually matter much
since you don't really need to do anything with it, assuming error
is not `null`, because that means the export was successful and you can
proceed by simply querying the resultant data in Seafowl.