Serialize the stepper state into the URL, so that awaiting can be resumed across page loads
Keep track of the current stepper state (e.g. taskId, import completion, etc.)
in the URL. Update the URL when the state changes, and initialize the state
from the URL on page load. Note that we need to default to an "uninitialized"
state, and then update the state from the URL via an `initialize_from_url` action, because
the `useRouter` hook is ansynchronous, and we don't look at query parameters on
the server side with `getInitialProps` or similar. Thus we can show a loading bar
before showing the import form (or whatever we're showing based on the current state).

This makes development easier, since after a long import we can refresh the page
with the URL containing the task ID and start from there, rather than re-importing
every time. And it also makes it easier for users who can refresh the page without
losing progress if an import has already started (it will just poll the taskId from
the URL).