Track completed import/exports in a meta repository on Splitgraph
After an import/export has completed, insert a row into the meta table,
which we will also use to fetch the previously imported repositories
from the client side when rendering the sidebar. We don't have transactional
guarantees on the DDN, so we can't do `INSERT ON CONFLICT`, so instead we
avoid duplicate rows by first selecting the existing row, and returning `204`
if it's already been inserted into the `completed_repositories` table.

However, I did notice that when I inserted the same row twice, it only
showed up once when I made a selection in the Console. I don't know if
this was due to a race condition, a bug, or because it's using the entire
row as a compound primary key and for some reason requiring that it be unique.