Fix test that was failing because Splitgraph export job completed too quickly
The test defers an export task, and then checks its status, which
it expects to still be pending. Previously, a query exporting 5 rows
to parquet was completing quickly enough that the test would sometimes
fail because it expected the task to still be pending but it was complete.
Export a query of 10,000 random rows instead, to make it take a bit longer...