Update `useSql` hook to accept query building function and abort signal
If a function is passed as the query parameter, instead of a string,
then call it to create the query.

If an abort signal is passed, call it in the cleanup function of
the hook. If it's not passed, then create one by default and call that.

This solves an issue where first render could be empty if the query
wasn't ready yet (e.g. some interpolated value like table name came
from the URL which hasn't been parsed), and the abort signal implements
the correct behavior for unpredictable mounting sequences (you're not
supposed to rely on `useEffect` only executing once).